pyproject.toml
[project]
name = "support-agent"
version = "0.1.0"
description = "Whatsapp based support agent for customer query realted to products and details."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.7",
    "langgraph-checkpoint-postgres>=3.0.4",
    "pip>=26.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
